Feat: 마일리지 카테고리 모달 완성 + 컴포넌트 분리(#7)
